% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spork.R
\name{as.expression.plotmath}
\alias{as.expression.plotmath}
\title{Coerce Plotmath to Expression}
\usage{
\method{as.expression}{plotmath}(x, ...)
}
\arguments{
\item{x}{plotmath}

\item{...}{ignored arguments}
}
\value{
expression
}
\description{
Coerces plotmath to expression by parsing as text.
}
\examples{
x <- c(
  'V_c./F',
  'AUC_ss',
  'C_max_ss',
  'var^eta_j'
)
x <- as_spork(x)
x <- as_plotmath(x)
x
as.expression(x)[[4]]
as.expression(x[[4]])
class(as.expression(x))
lapply(as.expression(x), class)
as.expression(as_plotmath(as_spork('V_c./F')))
as.expression(as_plotmath(as_spork(character(0))))
library(magrittr)
'gravitational force (kg\\\\.m/s^2.)' \%>\%
  as_spork \%>\%
  as_plotmath \%>\%
  as.expression -> label
  label
}
\seealso{
Other spork: \code{\link{as_spork.character}},
  \code{\link{as_spork.factor}}, \code{\link{as_spork}}
}
\concept{spork}
\keyword{internal}
