% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spork.R
\name{as.list.spork}
\alias{as.list.spork}
\title{Coerce Spork to List}
\usage{
\method{as.list}{spork}(x, ...)
}
\arguments{
\item{x}{spork}

\item{...}{ignored}
}
\value{
list of spork
}
\description{
Coerces spork to list.  Each element inherits class.
Supports use of \code{\link{lapply}}.
}
\examples{
x <- as_spork(letters[1:5])
lapply(x, class)
}
\keyword{internal}
