% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spork.R
\name{as_greek.default}
\alias{as_greek.default}
\title{Convert character to Greek}
\usage{
\method{as_greek}{default}(x, ...)
}
\arguments{
\item{x}{coerced to character}

\item{...}{passed arguments}
}
\value{
class greek (character)
}
\description{
Converts character to Greek by assigning class 'greek'.
}
\examples{
as_greek(greek())

}
\seealso{
Other greek: 
\code{\link{as_greek}()}
}
\concept{greek}
\keyword{internal}
