% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spork.R
\name{as_greek}
\alias{as_greek}
\title{Convert to Greek}
\usage{
as_greek(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arguments}
}
\description{
Converts to Greek.  Generic, 
with method for character.
}
\seealso{
Other greek: 
\code{\link{as_greek.default}()}
}
\concept{greek}
\keyword{internal}
