% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{as_html.default}
\alias{as_html.default}
\title{Convert Default to Html}
\usage{
\method{as_html}{default}(x, ...)
}
\arguments{
\item{x}{inherits character}

\item{...}{passed arguments}
}
\value{
html
}
\description{
Coerces to spork, then to html.
}
\examples{
as_html('anything')
}
\seealso{
Other html: 
\code{\link{[.html}()},
\code{\link{[[.html}()},
\code{\link{as_html.greek}()},
\code{\link{as_html.spar}()},
\code{\link{as_html.spork}()},
\code{\link{as_html}()},
\code{\link{html2xml}()},
\code{\link{htmlToken}()}
}
\concept{html}
\keyword{internal}
