% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{as_html.greek}
\alias{as_html.greek}
\title{Convert Greek to Html}
\usage{
\method{as_html}{greek}(x, ...)
}
\arguments{
\item{x}{greek}

\item{...}{ignored}
}
\value{
html
}
\description{
Converts Greek letter names to html.
}
\examples{
as_html(greek())
}
\seealso{
Other html: 
\code{\link{[.html}()},
\code{\link{[[.html}()},
\code{\link{as_html.spar}()},
\code{\link{as_html.spork}()},
\code{\link{as_html}()},
\code{\link{html2xml}()},
\code{\link{htmlToken}()}
}
\concept{html}
\keyword{internal}
