% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{as_html}
\alias{as_html}
\title{Coerce to Html}
\usage{
as_html(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
html
}
\description{
Coerce to html.  Generic, with method
\code{\link{as_html.spork}}.
}
\examples{
example(as_html.spork)
}
\seealso{
Other html: 
\code{\link{[.html}()},
\code{\link{[[.html}()},
\code{\link{as_html.spar}()},
\code{\link{as_html.spork}()},
\code{\link{html2xml}()},
\code{\link{htmlToken}()}
}
\concept{html}
\keyword{internal}
