% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{as_latex.default}
\alias{as_latex.default}
\title{Convert Default to Latex}
\usage{
\method{as_latex}{default}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{ignored}
}
\value{
latex
}
\description{
Converts anything to latex by adding class 'latex'.
}
\examples{
as_latex('text')
}
\seealso{
Other latex: 
\code{\link{[.latex}()},
\code{\link{[[.latex}()},
\code{\link{as_latex.spar}()},
\code{\link{as_latex.spork}()},
\code{\link{as_latex}()},
\code{\link{concatenate.latex}()},
\code{\link{latexToken}()}

Other spork: 
\code{\link{[.spork}()},
\code{\link{[[.spork}()},
\code{\link{as.list.spork}()},
\code{\link{as.png.spork}()},
\code{\link{as_html.spork}()},
\code{\link{as_latex.spork}()},
\code{\link{as_plotmath.spork}()},
\code{\link{as_previews.spork}()},
\code{\link{as_spar.spork}()},
\code{\link{as_spork.character}()},
\code{\link{as_spork.factor}()},
\code{\link{as_spork.spork}()},
\code{\link{as_spork}()},
\code{\link{ggplot.spork}()}

Other interface: 
\code{\link{as.expression.plotmath}()},
\code{\link{as_html.spar}()},
\code{\link{as_html.spork}()},
\code{\link{as_latex.spar}()},
\code{\link{as_latex.spork}()},
\code{\link{as_plotmath.spar}()},
\code{\link{as_plotmath.spork}()},
\code{\link{as_previews.spork}()},
\code{\link{as_spork.character}()},
\code{\link{htmlToken}()},
\code{\link{latexToken}()},
\code{\link{plotmathToken}()}
}
\concept{interface}
\concept{latex}
\concept{spork}
