% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{as_latex.greek}
\alias{as_latex.greek}
\title{Convert Greek to Latex}
\usage{
\method{as_latex}{greek}(x, ...)
}
\arguments{
\item{x}{greek}

\item{...}{ignored}
}
\value{
latex
}
\description{
Converts Greek letter names to html.
}
\examples{
as_latex(greek())
}
\seealso{
Other latex: 
\code{\link{[.latex}()},
\code{\link{[[.latex}()},
\code{\link{as_latex.spar}()},
\code{\link{as_latex.spork}()},
\code{\link{as_latex}()},
\code{\link{concatenate.latex}()},
\code{\link{latexToken}()}
}
\concept{latex}
\keyword{internal}
