% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{as_latex.latex}
\alias{as_latex.latex}
\title{Coerce Latex to Latex}
\usage{
\method{as_latex}{latex}(x, ...)
}
\arguments{
\item{x}{latex}

\item{...}{ignored}
}
\value{
latex
}
\description{
Coerces class 'latex' to 'latex'.  Simply returns x.
}
\examples{
class(as_latex(as_latex(letters)))
}
\seealso{
Other latex: 
\code{\link{[.latex}()},
\code{\link{[[.latex}()},
\code{\link{as_latex}()},
\code{\link{as_latex.default}()},
\code{\link{as_latex.greek}()},
\code{\link{as_latex.spar}()},
\code{\link{as_latex.spork}()},
\code{\link{concatenate.latex}()},
\code{\link{latexToken}()}
}
\concept{latex}
\keyword{internal}
