% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{as_latex}
\alias{as_latex}
\title{Coerce to Latex}
\usage{
as_latex(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
latex
}
\description{
Coerce to latex.  Generic, with method
\code{\link{as_latex.spork}}.
}
\examples{
example(as_latex.spork)
}
\seealso{
Other latex: 
\code{\link{as_latex.spar}()},
\code{\link{as_latex.spork}()},
\code{\link{latexToken}()}
}
\concept{latex}
\keyword{internal}
