% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{as_latex.spar}
\alias{as_latex.spar}
\title{Convert One Spork to Latex}
\usage{
\method{as_latex}{spar}(
  x,
  unrecognized = getOption("latex_unrecognized", "latexToken"),
  italics = FALSE,
  math = TRUE,
  ...
)
}
\arguments{
\item{x}{character}

\item{unrecognized}{function to process unrecognized tokens: default \code{\link{latexToken}}}

\item{italics}{whether to use italics or not (default: no)}

\item{math}{whether to wrap in math environment (default: yes)}

\item{...}{passed to \code{unrecognized}; see \code{\link{latexToken}}}
}
\value{
character
}
\description{
Converts one spork to latex.
See description for \code{\link{as_spork}}.
By default, unrecognized tokens are returned
literally.  However, Greek symbols and latex
metacharacters are escaped.
See \code{\link{latexToken}}.
}
\examples{
library(magrittr)
'V_c./F' \%>\% as_spork \%>\% as_latex
'AUC_ss' \%>\% as_spork \%>\% as_latex
'C_max_ss' \%>\% as_spork \%>\% as_latex
'var^eta_j' \%>\% as_spork \%>\% as_latex
'& \% $ # \\\\_ { } ~ \\\\^ \\\\' \%>\% as_spork \%>\% as_latex
'one joule (Omega) ~ 1 kg*m^2./s^2' \%>\% as_spork \%>\% as_latex
}
\seealso{
Other interface: 
\code{\link{as.expression.plotmath}()},
\code{\link{as_latex.spork}()},
\code{\link{as_plotmath.spar}()},
\code{\link{as_plotmath.spork}()},
\code{\link{as_spork.character}()},
\code{\link{concatenate.character}()},
\code{\link{concatenate.latex}()},
\code{\link{concatenate.plotmath}()},
\code{\link{latexToken}()},
\code{\link{plotmathToken}()}

Other latex: 
\code{\link{as_latex.spork}()},
\code{\link{as_latex}()},
\code{\link{latexToken}()}
}
\concept{interface}
\concept{latex}
