% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmath.R
\name{as_plotmath.default}
\alias{as_plotmath.default}
\title{Convert Default to Plotmath}
\usage{
\method{as_plotmath}{default}(x, ...)
}
\arguments{
\item{x}{inherits character}

\item{...}{passed arguments}
}
\value{
plotmath
}
\description{
Coerces to spork, then to plotmath.
}
\examples{
as_plotmath('anything')
}
\seealso{
Other plotmath: 
\code{\link{[.plotmath}()},
\code{\link{[[.plotmath}()},
\code{\link{as.expression.plotmath}()},
\code{\link{as.png.plotmath}()},
\code{\link{as_plotmath.greek}()},
\code{\link{as_plotmath.spar}()},
\code{\link{as_plotmath.spork}()},
\code{\link{as_plotmath}()},
\code{\link{concatenate.plotmath}()},
\code{\link{ggplot.plotmath}()},
\code{\link{goodToken}()},
\code{\link{plotmathToken}()}
}
\concept{plotmath}
\keyword{internal}
