% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmath.R
\name{as_plotmath.greek}
\alias{as_plotmath.greek}
\title{Convert Greek to Plotmath}
\usage{
\method{as_plotmath}{greek}(x, ...)
}
\arguments{
\item{x}{greek}

\item{...}{ignored}
}
\value{
plotmath
}
\description{
Converts Greek letter names to plotmath.
}
\examples{
as_plotmath(greek())
}
\seealso{
Other plotmath: 
\code{\link{[.plotmath}()},
\code{\link{[[.plotmath}()},
\code{\link{as.expression.plotmath}()},
\code{\link{as.png.plotmath}()},
\code{\link{as_plotmath.default}()},
\code{\link{as_plotmath.spar}()},
\code{\link{as_plotmath.spork}()},
\code{\link{as_plotmath}()},
\code{\link{concatenate.plotmath}()},
\code{\link{ggplot.plotmath}()},
\code{\link{goodToken}()},
\code{\link{plotmathToken}()}
}
\concept{plotmath}
\keyword{internal}
