% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spork.R
\name{as_plotmath}
\alias{as_plotmath}
\title{Coerce to Plotmath}
\usage{
as_plotmath(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
plotmath
}
\description{
Coerce to plotmath.  Generic, with method
\code{\link{as_plotmath.spork}}.
}
\examples{
example(as_plotmath.spork)
}
\seealso{
Other plotmath: 
\code{\link{as_plotmath.spork}()},
\code{\link{goodToken}()},
\code{\link{plotmathToken}()},
\code{\link{spork_to_plotmath}()}
}
\concept{plotmath}
\keyword{internal}
