% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmath.R
\name{as_plotmath.spar}
\alias{as_plotmath.spar}
\title{Convert One Spork to Plotmath}
\usage{
\method{as_plotmath}{spar}(
  x,
  unrecognized = getOption("plotmath_unrecognized", "plotmathToken"),
  ...
)
}
\arguments{
\item{x}{character}

\item{unrecognized}{function to process unrecognized tokens: default \code{\link{plotmathToken}}}

\item{...}{passed to \code{unrecognized}; see \code{\link{plotmathToken}}}
}
\value{
character
}
\description{
Converts one spork to plotmath.
See description for \code{\link{as_spork}}.
By default, unrecognized tokens are returned
unmodified if they are parseable.
Otherwise, backslashes and single quotes are escaped,
and the result is wrapped in single quotes.
See \code{\link{plotmathToken}}.
}
\examples{
library(magrittr)
'V_c./F' \%>\% as_spork \%>\% as_plotmath
'AUC_ss' \%>\% as_spork \%>\% as_plotmath
'C_max_ss' \%>\% as_spork \%>\% as_plotmath
'var^eta_j' \%>\% as_spork \%>\% as_plotmath
}
\seealso{
Other interface: 
\code{\link{as.expression.plotmath}()},
\code{\link{as_latex.spar}()},
\code{\link{as_latex.spork}()},
\code{\link{as_plotmath.spork}()},
\code{\link{as_spork.character}()},
\code{\link{concatenate.character}()},
\code{\link{concatenate.latex}()},
\code{\link{concatenate.plotmath}()},
\code{\link{latexToken}()},
\code{\link{plotmathToken}()}

Other plotmath: 
\code{\link{as_plotmath.spork}()},
\code{\link{as_plotmath}()},
\code{\link{goodToken}()},
\code{\link{plotmathToken}()}
}
\concept{interface}
\concept{plotmath}
