% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spork.R
\name{as_plotmath.spork}
\alias{as_plotmath.spork}
\title{Convert Spork to Plotmath}
\usage{
\method{as_plotmath}{spork}(x, ...)
}
\arguments{
\item{x}{spork}

\item{...}{ignored}
}
\value{
plotmath
}
\description{
Converts spork to plotmath. See \code{\link[grDevices]{plotmath}}.
Vectorized version of \code{\link{spork_to_plotmath}}.
}
\examples{
library(magrittr)
'V_c./F' \%>\% as_spork \%>\% as_plotmath
'AUC_ss' \%>\% as_spork \%>\% as_plotmath
'C_max_ss' \%>\% as_spork \%>\% as_plotmath
'var^eta_j' \%>\% as_spork \%>\% as_plotmath
'one joule (Omega) ~ 1 kg*m^2./s^2' \%>\% as_spork \%>\% as_plotmath
}
\seealso{
Other plotmath: \code{\link{as_plotmath}},
  \code{\link{goodToken}}, \code{\link{plotmathToken}},
  \code{\link{spork_to_plotmath}}
}
\concept{plotmath}
