% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{as_preview.html}
\alias{as_preview.html}
\title{Preview Spork as Html}
\usage{
\method{as_preview}{html}(x, ...)
}
\arguments{
\item{x}{spork; see \code{\link{as_spork}}}

\item{...}{ignored arguments}
}
\value{
kableExtra
}
\description{
Preview spork after conversion to html
Invokes the viewer using the print method for class 'kableExtra'.
}
\examples{
library(magrittr)
'one joule (Omega) ~ 1 kg*m^2./s^2' \%>\% 
as_html \%>\% 
as_preview
}
\seealso{
Other preview: 
\code{\link{as.png.plotmath}()},
\code{\link{as.png.spork}()},
\code{\link{as_preview.latex}()},
\code{\link{as_preview.plotmath}()},
\code{\link{as_previews.default}()},
\code{\link{as_previews.spork}()},
\code{\link{as_previews}()},
\code{\link{as_preview}()},
\code{\link{ggplot.plotmath}()},
\code{\link{ggplot.spork}()}
}
\concept{preview}
\keyword{internal}
