% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{as_preview.plotmath}
\alias{as_preview.plotmath}
\title{Preview Spork as Plotmath}
\usage{
\method{as_preview}{plotmath}(x, stem = "plotmath_preview", width = 3, height = 1, ...)
}
\arguments{
\item{x}{spork; see \code{\link{as_spork}}}

\item{width}{passed to \code{\link{as.png.plotmath}}}

\item{height}{passed to \code{\link{as.png.plotmath}}}

\item{...}{passed arguments}
}
\value{
invisible filepath
}
\description{
Preview spork after conversion to plotmath.
Creates and displays a temporary png file with
a parsed expression.
}
\examples{
library(magrittr)
'one joule (Omega) ~ 1 kg*m^2./s^2' \%>\%
as_spork \%>\%
as_plotmath \%>\%
as_preview
}
\seealso{
Other preview: 
\code{\link{as.png.plotmath}()},
\code{\link{as.png.spork}()},
\code{\link{as_preview.latex}()},
\code{\link{as_previews.spork}()},
\code{\link{as_previews}()},
\code{\link{as_preview}()},
\code{\link{ggplot.plotmath}()},
\code{\link{ggplot.spork}()}
}
\concept{preview}
\keyword{internal}
