% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{as_preview}
\alias{as_preview}
\title{Preview Something}
\usage{
as_preview(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Creates a preview.
Generic, with methods for latex and plotmath.
}
\examples{
library(magrittr)
'V_c./F' \%>\% as_spork \%>\% as_plotmath \%>\% as_preview
\donttest{
'V_c./F' \%>\% as_spork \%>\% as_latex \%>\% as_preview
}
'one joule (Omega) ~ 1 kg*m^2./s^2' \%>\% as_spork \%>\% as_plotmath \%>\% as_preview
\donttest{
'one joule (Omega) ~ 1 kg*m^2./s^2' \%>\% as_spork \%>\% as_latex \%>\% as_preview
}
}
\seealso{
Other preview: 
\code{\link{as.png.plotmath}()},
\code{\link{as.png.spork}()},
\code{\link{as_preview.latex}()},
\code{\link{as_preview.plotmath}()},
\code{\link{as_previews.spork}()},
\code{\link{as_previews}()},
\code{\link{ggplot.plotmath}()},
\code{\link{ggplot.spork}()}
}
\concept{preview}
\keyword{internal}
