% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{as_previews.default}
\alias{as_previews.default}
\title{Default Compare Previews}
\usage{
\method{as_previews}{default}(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{passed arguments}
}
\value{
see as_previews.spork
}
\description{
Compare previews by default assuming input is spork..
}
\examples{
example(as_previews.spork)
}
\seealso{
Other preview: 
\code{\link{as.png.plotmath}()},
\code{\link{as.png.spork}()},
\code{\link{as_preview}()},
\code{\link{as_preview.html}()},
\code{\link{as_preview.latex}()},
\code{\link{as_preview.plotmath}()},
\code{\link{as_previews}()},
\code{\link{as_previews.spork}()},
\code{\link{ggplot.plotmath}()},
\code{\link{ggplot.spork}()}
}
\concept{preview}
\keyword{internal}
