% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{as_previews}
\alias{as_previews}
\title{Compare Previews}
\usage{
as_previews(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Compare previews of something.
Generic, with method \code{\link{as_previews.spork}}.
}
\examples{
example(as_previews.spork)
}
\seealso{
Other preview: 
\code{\link{as.png.plotmath}()},
\code{\link{as.png.spork}()},
\code{\link{as_preview}()},
\code{\link{as_preview.html}()},
\code{\link{as_preview.latex}()},
\code{\link{as_preview.plotmath}()},
\code{\link{as_previews.default}()},
\code{\link{as_previews.spork}()},
\code{\link{ggplot.plotmath}()},
\code{\link{ggplot.spork}()}
}
\concept{preview}
\keyword{internal}
