% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spar.R
\name{as_spar.default}
\alias{as_spar.default}
\title{Parse Spork by Default}
\usage{
\method{as_spar}{default}(x, ...)
}
\arguments{
\item{x}{length-one character using spork syntax}

\item{...}{ignored arguments}
}
\value{
spar (character vector)
}
\description{
Parses spork by default. Coerces input with \code{as_spork()}.
}
\seealso{
Other spar: 
\code{\link{as_plotmath.spar}()},
\code{\link{as_spar.spork}()},
\code{\link{as_spar}()}

Other spork: 
\code{\link{[.spork}()},
\code{\link{[[.spork}()},
\code{\link{as.list.spork}()},
\code{\link{as.png.spork}()},
\code{\link{as_html.spork}()},
\code{\link{as_latex.spork}()},
\code{\link{as_plotmath.spork}()},
\code{\link{as_previews.spork}()},
\code{\link{as_spar.spork}()},
\code{\link{as_spork.character}()},
\code{\link{as_spork.factor}()},
\code{\link{as_spork.spork}()},
\code{\link{as_spork}()},
\code{\link{ggplot.spork}()},
\code{\link{greek}()}
}
\concept{spar}
\concept{spork}
\keyword{internal}
