% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spar.R
\name{as_spar}
\alias{as_spar}
\title{Parse Spork.}
\usage{
as_spar(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Parses spork.
Generic, with method \code{\link{as_spar.spork}}.
}
\examples{
# see methods
}
\seealso{
Other spar: 
\code{\link{as_plotmath.spar}()},
\code{\link{as_spar.default}()},
\code{\link{as_spar.spork}()}
}
\concept{generics}
\concept{spar}
\keyword{internal}
