% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spork.R
\name{as_spork.character}
\alias{as_spork.character}
\title{Coerce Character to Spork}
\usage{
\method{as_spork}{character}(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{ignored arguments}
}
\value{
spork
}
\description{
Coerces character to class 'spork'.
See description for \code{\link{as_spork}}.
}
\examples{
as_spork('V_c./F')
}
\seealso{
Other spork: \code{\link{as.expression.plotmath}},
  \code{\link{as_spork.factor}}, \code{\link{as_spork}}
}
\concept{spork}
