% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spork.R
\name{as_spork.factor}
\alias{as_spork.factor}
\title{Coerce Factor to Spork}
\usage{
\method{as_spork}{factor}(x, ...)
}
\arguments{
\item{x}{factor}

\item{...}{ignored arguments}
}
\value{
spork
}
\description{
Coerces factor to class 'spork'
by converting to character and calling
\code{\link{as_spork}}.
}
\examples{
as_spork(as.factor('V_c./F'))
}
\seealso{
Other spork: 
\code{\link{as.expression.plotmath}()},
\code{\link{as_spork.character}()},
\code{\link{as_spork}()}
}
\concept{spork}
\keyword{internal}
