% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spork.R
\name{as_spork}
\alias{as_spork}
\title{Coerce to Spork}
\usage{
as_spork(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
spork
}
\description{
Coerces to class 'spork'. Generic,
with method \code{\link{as_spork.character}}.
A spork is simple text expressing
arbitrarily nested subscripts (\code{x_y_z})
and superscripts (\code{x^y^z}). A dot
(\code{x^y._z}) explicitly terminates
a group. An asterisk (\code{*}) suggests
multiplication. Special characters
may be escaped with a backslash.
Convert to plotmath with\code{\link{as_plotmath}}
and to latex with \code{\link{as_latex}}.
}
\examples{
library(magrittr)
library(dplyr)
library(latexpdf)

'j^*' \%>\% as_spork \%>\% as_plotmath \%>\% as.character
as_spork \%>\%
as_plotmath #\%>\%
as_expression
data.frame(y=1:10, x=1:10) \%>\%
decorate("x: one joule (Omega) ~ 1 kg*m^2./s^2") \%>\%
mutate(
 x = structure(x, label = x \%>\% attr('label') \%>\%
 as_spork \%>\%
 as_plotmath \%>\%
 as.expression)
) \%>\%
 ggplot(aes(x, y))

data.frame(y=1:10, x=1:10) \%>\%
decorate("x: gravitational force - gamma (kg\\\\.m/s^2.)") \%>\%
mutate(x = structure(x, label = x \%>\% attr('label') \%>\%
as_spork \%>\%
as_plotmath \%>\%
as.expression)) \%>\%
ggplot(aes(x, y))

path <- tempfile()
data.frame(
 stringsAsFactors = FALSE,
 spork = c(
   'one joule (Omega) ~ 1 kg*m^2./s^2',
   'gravitational force - gamma (kg\\\\.m/s^2.)'
 )
) \%>\%
 transmute(
   latex = spork \%>\%
   as_spork \%>\%
   as_latex
 ) \%>\%
 as.pdf(reserve = FALSE, wider = -70)
}
\seealso{
Other spork: 
\code{\link{as.expression.plotmath}()},
\code{\link{as_spork.character}()},
\code{\link{as_spork.factor}()},
\code{\link{as_spork.unit_string}()},
\code{\link{as_spork.units}()}
}
\concept{spork}
\keyword{internal}
