% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spork.R
\name{as_spork.spork}
\alias{as_spork.spork}
\title{Coerce Spork to Spork}
\usage{
\method{as_spork}{spork}(x, ...)
}
\arguments{
\item{x}{spork}

\item{...}{ignored arguments}
}
\value{
spork
}
\description{
Coerces 'spork' to class 'spork'.
Supplies any implied terminal dots.
}
\examples{
as_spork(as_spork(c('C_max_ss','t^1/2','V_c./F','foo')))
as_spork(as_spork('V_c./F'))
as_spork(as_spork('V_c'))
}
\seealso{
Other spork: 
\code{\link{[.spork}()},
\code{\link{[[.spork}()},
\code{\link{as.list.spork}()},
\code{\link{as.png.spork}()},
\code{\link{as_html.spork}()},
\code{\link{as_latex.spork}()},
\code{\link{as_plotmath.spork}()},
\code{\link{as_previews.spork}()},
\code{\link{as_spar.spork}()},
\code{\link{as_spork.character}()},
\code{\link{as_spork.factor}()},
\code{\link{as_spork}()},
\code{\link{ggplot.spork}()}
}
\concept{spork}
\keyword{internal}
