% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate.R
\name{concatenate.NULL}
\alias{concatenate.NULL}
\title{Concatenate NULL}
\usage{
\method{concatenate}{`NULL`}(x, ...)
}
\arguments{
\item{x}{NULL}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Concatenates null. Returns empty string.
}
\examples{
concatenate(NULL)
}
\seealso{
Other concatenate: 
\code{\link{concatenate}()},
\code{\link{concatenate.character}()},
\code{\link{concatenate.latex}()},
\code{\link{concatenate.plotmath}()}
}
\concept{concatenate}
\keyword{internal}
