% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate.R
\name{concatenate.plotmath}
\alias{concatenate.plotmath}
\title{Concatenate Plotmath}
\usage{
\method{concatenate}{plotmath}(x, sep = "*", ...)
}
\arguments{
\item{x}{plotmath}

\item{sep}{character}

\item{...}{passed arguments}
}
\value{
plotmath
}
\description{
Concatenates plotmath. Collapses vector using separator.
}
\examples{
concatenate(as_plotmath(as_spork('BMI_i','kg/m^2')))
}
\seealso{
Other interface: 
\code{\link{as.expression.plotmath}()},
\code{\link{as_latex.spork}()},
\code{\link{as_plotmath.spork}()},
\code{\link{as_spork.character}()},
\code{\link{concatenate.character}()},
\code{\link{concatenate.latex}()},
\code{\link{latexToken}()},
\code{\link{plotmathToken}()},
\code{\link{spork_to_latex}()},
\code{\link{spork_to_plotmath}()}

Other concatenate: 
\code{\link{concatenate.NULL}()},
\code{\link{concatenate.character}()},
\code{\link{concatenate.latex}()},
\code{\link{concatenate}()}
}
\concept{concatenate}
\concept{interface}
