% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate.R
\name{concatenate}
\alias{concatenate}
\title{Concatenate Something}
\usage{
concatenate(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Concatenates something.
Generic, with methods
\code{\link{concatenate.character}},
\code{\link{concatenate.plotmath}}, and
\code{\link{concatenate.latex}}.
}
\examples{
# see methods
}
\seealso{
Other concatenate: \code{\link{concatenate.NULL}},
  \code{\link{concatenate.character}},
  \code{\link{concatenate.latex}},
  \code{\link{concatenate.plotmath}}
}
\concept{concatenate}
\keyword{internal}
