% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explicit.R
\name{explicit}
\alias{explicit}
\title{Make Something Explicit}
\usage{
explicit(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
see methods
}
\description{
Makes something explicit.
Generic, with method
\code{\link{explicit.spork}}.
}
\examples{
# see methods
}
\seealso{
Other explicit: 
\code{\link{explicit.spork}()}
}
\concept{explicit}
\keyword{internal}
