% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explicit.R
\name{explicit.spork}
\alias{explicit.spork}
\title{Make Spork Explicit}
\usage{
\method{explicit}{spork}(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{passed arguments}
}
\value{
spork
}
\description{
Makes spork explicit. For every section
as indicated by line breaks, it supplies
dots to close open scripts.
Explicit spork can be concatenated more easily.
One generally wants to avoid breaking lines
within active sub- and super scripts.
}
\examples{
s <- c(
  'Work_out\\\\nx^y\\\\n',
  '\\\\nkg/m^2/S^2'
)
explicit(as_spork(s))
}
\seealso{
Other explicit: 
\code{\link{explicit}()}
}
\concept{explicit}
\keyword{internal}
