% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatNumeric.R
\name{formatNumeric}
\alias{formatNumeric}
\title{Convert a numeric vector to character while applying pretty formatting}
\usage{
formatNumeric(x, type, digits = 0)
}
\arguments{
\item{x}{numeric vector on which to apply formatting/conversion}

\item{type}{the type of formatting to use; valid values are "int", "dlr" and 
"pct"}

\item{digits}{non-negative integer value indicating the number of decimal 
places to use}
}
\value{
A character vector of the converted/formatted numeric input.
}
\description{
\code{formatNumeric} converts numeric values to character 
and can apply various pretty formatting
}
\examples{
x <- rnorm(5)
print(x)
formatNumeric(x, "pct", 1)
formatNumeric(x*10000, "int")
formatNumeric(x*10000, "dlr", digits = 2)
}
\seealso{
\code{\link{round}, \link{trimws}}
}

