% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{ggplot.plotmath}
\alias{ggplot.plotmath}
\title{Plot Plotmath}
\usage{
\method{ggplot}{plotmath}(x, blank = TRUE, ...)
}
\arguments{
\item{x}{length-one plotmath; see \code{\link{as_plotmath}}}

\item{blank}{whether to use a blank plot area}

\item{...}{ignored arguments}
}
\value{
gg
}
\description{
Render plotmath in a ggplot.
}
\examples{
library(magrittr)
'one joule (Omega) ~ 1 kg*m^2./s^2' \%>\% as_spork \%>\% as_plotmath \%>\% ggplot
}
\seealso{
Other preview: 
\code{\link{as.png.plotmath}()},
\code{\link{as.png.spork}()},
\code{\link{as_preview.latex}()},
\code{\link{as_preview.plotmath}()},
\code{\link{as_previews.spork}()},
\code{\link{as_previews}()},
\code{\link{as_preview}()},
\code{\link{ggplot.spork}()}
}
\concept{preview}
\keyword{internal}
