% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spork-to-plotmath.R
\name{goodToken}
\alias{goodToken}
\title{Test Whether Token is Parseable}
\usage{
goodToken(x, ...)
}
\arguments{
\item{x}{length-one character}

\item{...}{ignored arguments}
}
\value{
logical
}
\description{
Tests whether token is Parseable
}
\examples{
goodToken('alpha')
goodToken('foo')
goodToken('\\\\$')
}
\seealso{
Other plotmath: \code{\link{as_plotmath.spork}},
  \code{\link{as_plotmath}}, \code{\link{plotmathToken}},
  \code{\link{spork_to_plotmath}}
}
\concept{plotmath}
\keyword{internal}
