% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spork.R
\name{greek}
\alias{greek}
\title{Greek letters recognized by Spork}
\usage{
greek()
}
\value{
class 'greek' (character)
}
\description{
Returns words representing the Greek
letters recognized by \pkg{spork}.
}
\details{
See \code{\link{plotmath}} for related detail.
In addition to the 24 uppercase and 24
lowercase Greek letters, 
latex supports 
varepsilon, vartheta, varpi, varsigma, and varrho
while plotmath supports
theta1, phi1, sigma1, omega1, Upsilon1, and stigma1.
Spork supports all, and tries to do something useful
for those not supported in html, latex, or plotmath.
}
\examples{
greek()
}
\seealso{
Other spork: 
\code{\link{[.spork}()},
\code{\link{[[.spork}()},
\code{\link{as.list.spork}()},
\code{\link{as.png.spork}()},
\code{\link{as_html.spork}()},
\code{\link{as_latex.spork}()},
\code{\link{as_plotmath.spork}()},
\code{\link{as_previews.spork}()},
\code{\link{as_spar.default}()},
\code{\link{as_spar.spork}()},
\code{\link{as_spork.character}()},
\code{\link{as_spork.factor}()},
\code{\link{as_spork.spork}()},
\code{\link{as_spork}()},
\code{\link{ggplot.spork}()}
}
\concept{spork}
\keyword{internal}
