% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html2xml}
\alias{html2xml}
\title{Convert HTML Greek entity references to XML}
\usage{
html2xml(x, ...)
}
\arguments{
\item{x}{html character}

\item{...}{ignored}
}
\value{
html
}
\description{
Converts HTML Greek entity references to XML entity references.
This is necessary because of a kableExtra bug: 
https://github.com/haozhu233/kableExtra/issues/814 .
The mappings used here are from 
https://www.thoughtco.com/html-codes-greek-characters-4062212
}
\examples{
htmlToken('alpha')
htmlToken('Upsilon1')
htmlToken('vartheta')
htmlToken('stigma')
htmlToken('varrho')
htmlToken('varpi')
htmlToken('Upsilon1')
}
\seealso{
Other html: 
\code{\link{[.html}()},
\code{\link{[[.html}()},
\code{\link{as_html}()},
\code{\link{as_html.greek}()},
\code{\link{as_html.spar}()},
\code{\link{as_html.spork}()},
\code{\link{htmlToken}()}
}
\concept{html}
\keyword{internal}
