% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{latexToken}
\alias{latexToken}
\title{Process Latex Token}
\usage{
latexToken(x, unrecognized = latexToken, math = TRUE, italics = FALSE, ...)
}
\arguments{
\item{x}{character}

\item{unrecognized}{function to process unrecognized tokens}

\item{math}{whether to wrap in math environment}

\item{italics}{whether to use italics or not}

\item{...}{ignored arguments}
}
\value{
latex
}
\description{
Pre-processes a latex token not recognized as
spork.  Escapes the common names for Greek letters
and escapes latex metacharacters.
}
\examples{
latexToken('foo')
latexToken('alpha')
latexToken('Alpha')
}
\seealso{
Other latex: 
\code{\link{as_latex.spar}()},
\code{\link{as_latex.spork}()},
\code{\link{as_latex}()}

Other interface: 
\code{\link{as.expression.plotmath}()},
\code{\link{as_latex.spar}()},
\code{\link{as_latex.spork}()},
\code{\link{as_plotmath.spar}()},
\code{\link{as_plotmath.spork}()},
\code{\link{as_spork.character}()},
\code{\link{concatenate.character}()},
\code{\link{concatenate.latex}()},
\code{\link{concatenate.plotmath}()},
\code{\link{plotmathToken}()}
}
\concept{interface}
\concept{latex}
