% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPackages.R
\name{loadPackages}
\alias{loadPackages}
\title{Load the packages specified in a file}
\usage{
loadPackages(f)
}
\arguments{
\item{f}{character value; path/name of the file containing the package list}
}
\value{
A logical vector describing the success of each \code{library} call.
}
\description{
\code{loadPackages} attempts to load all of the packages listed 
in the specified file
}
\examples{
writeLines(c("dplyr", "ggplot2", "rmarkdown"), "PACKAGES")
loadPackages("PACKAGES")
}
\seealso{
\code{\link{library}, \link{readLines}}
}

