% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDataCreate.R
\name{RDataCreate}
\alias{RDataCreate}
\title{Create an RData file containing the results of a function}
\usage{
RDataCreate(dir, x, fun, args, replace = FALSE)
}
\arguments{
\item{dir}{character value; directory in which to store the output RData file}

\item{x}{character value; name to use for the file/object}

\item{fun}{name of the function to call (non-quoted)}

\item{args}{named list of arguments to supply to the function; the 
\code{names} attribute of \code{args} gives the argument names}

\item{replace}{logical value; replace the output RData file if it already 
exists?}
}
\value{
The result of the evaluated function call is stored in an RData 
file using the path/name: \code{paste0(dir, "/", x, ".RData")}
}
\description{
\code{RDataCreate} stores the result of a function call in an 
RData file using the specified directory and file/object name
}
\examples{
RDataCreate(".", "test", names, list(x = mtcars), TRUE)
load("test.RData")
print(test)
}
\seealso{
\code{\link{RDataUse}, \link{do.call}, \link{assign}, \link{save}}
}

