% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDataUse.R
\name{RDataUse}
\alias{RDataUse}
\title{Return a specific RData file/object}
\usage{
RDataUse(dir, x)
}
\arguments{
\item{dir}{character value; path to the RData file}

\item{x}{character value; file/object name to retrieve}
}
\value{
Returns the object \code{x} from the RData file with the path/name: 
\code{paste0(dir, "/", x, ".RData")}
}
\description{
\code{RDataUse} returns the specified RData file/object. The 
file and object must have the same name.
}
\examples{
RDataCreate(".", "test", names, list(x = mtcars), TRUE)
load("test.RData")
print(test)
x <- RDataUse(".", "test")
print(x)
identical(test, x)
}
\seealso{
\code{\link{RDataCreate}, \link{new.env}, \link{load}, \link{get}}
}

