% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readConfig.R
\name{readConfig}
\alias{readConfig}
\title{Read a configuration file into a named list}
\usage{
readConfig(file)
}
\arguments{
\item{file}{character value; path/name of the configuration file}
}
\value{
A named list of character values where the names are the keys from 
the configuration file and the values are the values from the configuration 
file.
}
\description{
\code{readConfig} reads a configuration file of key value pairs 
and stores the result in a named list
}
\examples{
writeLines(c("COLOR=blue", "ANIMAL=dog"), "test")
readConfig("test")
}
\seealso{
\code{\link{readLines}, \link{strsplit}}
}

