% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceFunctions.R
\name{sourceFunctions}
\alias{sourceFunctions}
\title{Source a directory of R functions}
\usage{
sourceFunctions(dir, recursive = TRUE)
}
\arguments{
\item{dir}{path containing the files to source}

\item{recursive}{logical value; recurse into sub-directories?}
}
\value{
The \code{source} function is applied to any {*.R} files found in 
\code{dir}.
}
\description{
\code{sourceFunctions} sources all of the \code{*.R} files 
found in the specified directory
}
\examples{
write("foo <- function(){'Hello, world!'}", "foo.R")
sourceFunctions(".")
}
\seealso{
\code{\link{source}, \link{list.files}}
}

