% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sporklet.R
\name{sporklet}
\alias{sporklet}
\title{Parse Spork}
\usage{
sporklet(x, ...)
}
\arguments{
\item{x}{length-one character using spork syntax; coerced to character}

\item{...}{ignored arguments}
}
\value{
character
}
\description{
Parses spork.  Converts length-one character
to vector of tokens.  Explicit tokens include
\code{*._^} and any of these escaped with
backslash, e.g. \code{'\\*'}. One or more consecutive
whitespace characters are a single token,
as are one or more consecutive octothorpes (\code{#}).
Any string of characters delimited by
one or more of the above is implicitly
a token as well.
}
\examples{
sporklet('one joule (Omega) ~ 1 kg*m^2./s^2')
}
\concept{parse}
\keyword{internal}
