% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{[.latex}
\alias{[.latex}
\title{Subset Latex}
\usage{
\method{[}{latex}(x, ...)
}
\arguments{
\item{x}{latex}

\item{...}{passed to next method}
}
\value{
latex
}
\description{
Subsets latex, retaining class.
}
\examples{
x <- c(
  'V_c./F',
  'AUC_ss',
  'C_max_ss',
  'var^eta_j'
)
x <- as_latex(as_spork(x))
class(x)
class(x[1])
}
\seealso{
Other util: 
\code{\link{[.plotmath}()},
\code{\link{[.spork}()},
\code{\link{[[.latex}()},
\code{\link{[[.plotmath}()},
\code{\link{[[.spork}()}
}
\concept{util}
\keyword{internal}
