% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spork.R
\name{[.spork}
\alias{[.spork}
\title{Subset Spork}
\usage{
\method{[}{spork}(x, ...)
}
\arguments{
\item{x}{spork}

\item{...}{passed to next method}
}
\value{
spork
}
\description{
Subsets spork, retaining class.
}
\examples{
x <- c(
  'V_c./F',
  'AUC_ss',
  'C_max_ss',
  'var^eta_j'
)
x <- as_spork(x)
class(x)
class(x[1])
}
\seealso{
Other spork: \code{\link{[[.spork}},
  \code{\link{as.list.spork}}, \code{\link{as.png.spork}},
  \code{\link{as_latex.spork}},
  \code{\link{as_plotmath.spork}},
  \code{\link{as_previews.spork}},
  \code{\link{as_spar.spork}},
  \code{\link{as_spork.character}},
  \code{\link{as_spork.factor}},
  \code{\link{as_spork.spork}}, \code{\link{as_spork}},
  \code{\link{ggplot.spork}}
}
\concept{spork}
\keyword{internal}
