% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{[[.html}
\alias{[[.html}
\title{Element-select Html}
\usage{
\method{[[}{html}(x, ...)
}
\arguments{
\item{x}{html}

\item{...}{passed to next method}
}
\value{
html
}
\description{
Element-selects html, retaining class.
}
\examples{
x <- c(
  'V_c./F',
  'AUC_ss',
  'C_max_ss',
  'var^eta_j'
)
x <- as_html(as_spork(x))
class(x)
class(x[[1]])
}
\seealso{
Other html: 
\code{\link{[.html}()},
\code{\link{as_html.spar}()},
\code{\link{as_html.spork}()},
\code{\link{as_html}()},
\code{\link{htmlToken}()}
}
\concept{html}
\keyword{internal}
