% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{[[.latex}
\alias{[[.latex}
\title{Element-select Latex}
\usage{
\method{[[}{latex}(x, ...)
}
\arguments{
\item{x}{latex}

\item{...}{passed to next method}
}
\value{
latex
}
\description{
Element-selects latex, retaining class.
}
\examples{
x <- c(
  'V_c./F',
  'AUC_ss',
  'C_max_ss',
  'var^eta_j'
)
x <- as_latex(as_spork(x))
class(x)
class(x[[1]])
}
\seealso{
Other latex: 
\code{\link{[.latex}()},
\code{\link{as_latex.greek}()},
\code{\link{as_latex.spar}()},
\code{\link{as_latex.spork}()},
\code{\link{as_latex}()},
\code{\link{concatenate.latex}()},
\code{\link{latexToken}()}
}
\concept{latex}
\keyword{internal}
