% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmath.R
\name{[[.plotmath}
\alias{[[.plotmath}
\title{Element-select Plotmath}
\usage{
\method{[[}{plotmath}(x, ...)
}
\arguments{
\item{x}{plotmath}

\item{...}{passed to next method}
}
\value{
plotmath
}
\description{
Element-selects plotmath, retaining class.
}
\examples{
x <- c(
  'V_c./F',
  'AUC_ss',
  'C_max_ss',
  'var^eta_j'
)
x <- as_plotmath(as_spork(x))
class(x)
class(x[[1]])
}
\seealso{
Other plotmath: 
\code{\link{[.plotmath}()},
\code{\link{as.expression.plotmath}()},
\code{\link{as.png.plotmath}()},
\code{\link{as_plotmath.default}()},
\code{\link{as_plotmath.greek}()},
\code{\link{as_plotmath.spar}()},
\code{\link{as_plotmath.spork}()},
\code{\link{as_plotmath}()},
\code{\link{concatenate.plotmath}()},
\code{\link{ggplot.plotmath}()},
\code{\link{goodToken}()},
\code{\link{plotmathToken}()}
}
\concept{plotmath}
\keyword{internal}
