% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spork.R
\name{as.list.spork}
\alias{as.list.spork}
\title{Coerce Spork to List}
\usage{
\method{as.list}{spork}(x, ...)
}
\arguments{
\item{x}{spork}

\item{...}{ignored}
}
\value{
list of spork
}
\description{
Coerces spork to list.  Each element inherits class.
Supports use of \code{\link{lapply}}.
}
\examples{
x <- as_spork(letters[1:5])
lapply(x, class)
}
\seealso{
Other spork: 
\code{\link{[.spork}()},
\code{\link{[[.spork}()},
\code{\link{as.png.spork}()},
\code{\link{as_html.spork}()},
\code{\link{as_latex.default}()},
\code{\link{as_latex.spork}()},
\code{\link{as_plotmath.spork}()},
\code{\link{as_previews.spork}()},
\code{\link{as_spar.spork}()},
\code{\link{as_spork.character}()},
\code{\link{as_spork.factor}()},
\code{\link{as_spork.spork}()},
\code{\link{as_spork}()},
\code{\link{ggplot.spork}()}
}
\concept{spork}
\keyword{internal}
