% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{as_latex.default}
\alias{as_latex.default}
\title{Coerce Default to Latex}
\usage{
\method{as_latex}{default}(x, ...)
}
\arguments{
\item{x}{presumably inherits character}

\item{...}{ignored}
}
\value{
latex
}
\description{
Coerces non-latex to 'latex'. Simply prepends 'latex' to class path.
}
\examples{
class(as_latex(letters))
}
\seealso{
Other latex: 
\code{\link{[.latex}()},
\code{\link{[[.latex}()},
\code{\link{as_latex}()},
\code{\link{as_latex.greek}()},
\code{\link{as_latex.latex}()},
\code{\link{as_latex.spar}()},
\code{\link{as_latex.spork}()},
\code{\link{concatenate.latex}()},
\code{\link{latexToken}()}
}
\concept{latex}
\keyword{internal}
