% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{as_latex.greek}
\alias{as_latex.greek}
\title{Convert Greek to Latex}
\usage{
\method{as_latex}{greek}(x, ..., upgreek = getOption("spork_upgreek", TRUE))
}
\arguments{
\item{x}{greek}

\item{...}{ignored,}

\item{upgreek}{if TRUE, assume latex package \code{upgreek} is available}
}
\value{
latex
}
\description{
Converts Greek letter names to latex.
By default, assumes latex package \code{upgreek} is available
so that lower case greek can be typeset upright, e.g. using 
backslash-upalpha instead of backslash-alpha. Do use the package
"upgreek" in your document preamble,
or else in your R environment set \code{options(spork_upgreek = FALSE)}.
}
\examples{
as_latex(greek())
}
\seealso{
Other latex: 
\code{\link{[.latex}()},
\code{\link{[[.latex}()},
\code{\link{as_latex.spar}()},
\code{\link{as_latex.spork}()},
\code{\link{as_latex}()},
\code{\link{concatenate.latex}()},
\code{\link{latexToken}()}
}
\concept{latex}
\keyword{internal}
