% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spork.R
\name{as_latex.spork}
\alias{as_latex.spork}
\title{Convert Spork to Latex}
\usage{
\method{as_latex}{spork}(x, ...)
}
\arguments{
\item{x}{spork}

\item{...}{passed to \code{\link{spork_to_latex}}}
}
\value{
latex
}
\description{
Converts spork to latex.
Vectorized version of \code{\link{spork_to_latex}}.
}
\examples{
x <- c(
  'V_c./F',
  'AUC_ss',
  'C_max_ss',
  'var^eta_j'
)
x <- as_spork(x)
as_latex(x)
as_latex(as_spork('gravitational force (kg\\\\.m/s^2.)'))
}
\seealso{
Other latex: 
\code{\link{as_latex}()},
\code{\link{latexToken}()},
\code{\link{spork_to_latex}()}

Other interface: 
\code{\link{as.expression.plotmath}()},
\code{\link{as_plotmath.spork}()},
\code{\link{as_spork.character}()},
\code{\link{concatenate.character}()},
\code{\link{concatenate.latex}()},
\code{\link{concatenate.plotmath}()},
\code{\link{latexToken}()},
\code{\link{plotmathToken}()},
\code{\link{spork_to_latex}()},
\code{\link{spork_to_plotmath}()}
}
\concept{interface}
\concept{latex}
