% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{as_latex.spork}
\alias{as_latex.spork}
\title{Convert Spork to Latex}
\usage{
\method{as_latex}{spork}(x, ...)
}
\arguments{
\item{x}{spork}

\item{...}{passed to \code{\link{as_latex.spar}}}
}
\value{
latex
}
\description{
Converts spork to latex.
Vectorized version of \code{\link{as_latex.spar}}.
}
\examples{
x <- c(
  'V_c./F',
  '\\\\nAUC_ss',
  'C_max_ss\\\\n',
  'var^eta_j\\\\nrecords'
)
x <- as_spork(x)
writeLines(as_latex(x))
x <- as_spork('gravitational force\\\\n (kg\\\\.m/s^2.)')
explicit(x)
as_latex(x)
}
\seealso{
Other latex: 
\code{\link{[.latex}()},
\code{\link{[[.latex}()},
\code{\link{as_latex}()},
\code{\link{as_latex.default}()},
\code{\link{as_latex.greek}()},
\code{\link{as_latex.latex}()},
\code{\link{as_latex.spar}()},
\code{\link{concatenate.latex}()},
\code{\link{latexToken}()}

Other spork: 
\code{\link{[.spork}()},
\code{\link{[[.spork}()},
\code{\link{as.list.spork}()},
\code{\link{as.png.spork}()},
\code{\link{as_html.spork}()},
\code{\link{as_plotmath.spork}()},
\code{\link{as_previews.spork}()},
\code{\link{as_spar.default}()},
\code{\link{as_spar.spork}()},
\code{\link{as_spork}()},
\code{\link{as_spork.character}()},
\code{\link{as_spork.factor}()},
\code{\link{as_spork.spork}()},
\code{\link{ggplot.spork}()},
\code{\link{greek}()}

Other interface: 
\code{\link{as.expression.plotmath}()},
\code{\link{as_html.spar}()},
\code{\link{as_html.spork}()},
\code{\link{as_latex.spar}()},
\code{\link{as_plotmath.spar}()},
\code{\link{as_plotmath.spork}()},
\code{\link{as_previews.spork}()},
\code{\link{as_spork.character}()},
\code{\link{htmlToken}()},
\code{\link{latexToken}()},
\code{\link{plotmathToken}()}
}
\concept{interface}
\concept{latex}
\concept{spork}
