% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmath.R
\name{as_plotmath}
\alias{as_plotmath}
\title{Coerce to Plotmath}
\usage{
as_plotmath(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\value{
plotmath
}
\description{
Coerce to plotmath.  Generic, with method
\code{\link{as_plotmath.spork}}.
}
\examples{
example(as_plotmath.spork)
}
\seealso{
Other plotmath: 
\code{\link{[.plotmath}()},
\code{\link{[[.plotmath}()},
\code{\link{as.expression.plotmath}()},
\code{\link{as.png.plotmath}()},
\code{\link{as_plotmath.default}()},
\code{\link{as_plotmath.greek}()},
\code{\link{as_plotmath.spar}()},
\code{\link{as_plotmath.spork}()},
\code{\link{concatenate.plotmath}()},
\code{\link{ggplot.plotmath}()},
\code{\link{goodToken}()},
\code{\link{plotmathToken}()}
}
\concept{plotmath}
\keyword{internal}
