% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spork.R
\name{as_plotmath.spork}
\alias{as_plotmath.spork}
\title{Convert Spork to Plotmath}
\usage{
\method{as_plotmath}{spork}(x, ...)
}
\arguments{
\item{x}{spork}

\item{...}{ignored}
}
\value{
plotmath
}
\description{
Converts spork to plotmath. See '?plotmath'.
Vectorized version of \code{\link{spork_to_plotmath}}.
}
\examples{
library(magrittr)
'V_c./F' \%>\% as_plotmath
'AUC_ss' \%>\% as_plotmath
'C_max_ss' \%>\% as_plotmath
'var^eta_j' \%>\% as_plotmath
as_plotmath(as_spork('one joule (Omega) ~ 1 kg*m^2./s^2'))
}
\seealso{
Other plotmath: 
\code{\link{as_plotmath}()},
\code{\link{goodToken}()},
\code{\link{plotmathToken}()},
\code{\link{spork_to_plotmath}()}
}
\concept{plotmath}
