% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmath.R
\name{as_plotmath.spork}
\alias{as_plotmath.spork}
\title{Convert Spork to Plotmath}
\usage{
\method{as_plotmath}{spork}(x, ...)
}
\arguments{
\item{x}{spork}

\item{...}{passed to \code{\link{as_plotmath.spar}}}
}
\value{
plotmath
}
\description{
Converts spork to plotmath. See \code{\link[grDevices]{plotmath}}.
Vectorized version of \code{\link{as_plotmath.spar}}.
}
\examples{
library(magrittr)
'V_c./F' \%>\% as_spork \%>\% as_plotmath
'AUC_ss' \%>\% as_spork \%>\% as_plotmath
'C_max_ss' \%>\% as_spork \%>\% as_plotmath
'var^eta_j' \%>\% as_spork \%>\% as_plotmath
'one joule (Omega) ~ 1 kg*m^2./s^2' \%>\% as_spork \%>\% as_plotmath
}
\seealso{
Other plotmath: 
\code{\link{as_plotmath.spar}()},
\code{\link{as_plotmath}()},
\code{\link{goodToken}()},
\code{\link{plotmathToken}()}

Other interface: 
\code{\link{as.expression.plotmath}()},
\code{\link{as_latex.spar}()},
\code{\link{as_latex.spork}()},
\code{\link{as_plotmath.spar}()},
\code{\link{as_spork.character}()},
\code{\link{concatenate.character}()},
\code{\link{concatenate.latex}()},
\code{\link{concatenate.plotmath}()},
\code{\link{latexToken}()},
\code{\link{plotmathToken}()}
}
\concept{interface}
\concept{plotmath}
