% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{as_previews.spork}
\alias{as_previews.spork}
\title{Compare Previews of Spork}
\usage{
\method{as_previews}{spork}(x, wide = 70, long = 20, width = 3, height = 1, ...)
}
\arguments{
\item{x}{length-one spork}

\item{wide}{width in mm of the latex image}

\item{long}{length in mm of the latex image}

\item{width}{width (default: inches) of the plotmath image}

\item{height}{height (default: inches) of the plotmath image}

\item{...}{passed arguments}
}
\value{
invisible list of filepaths
}
\description{
Compares plotmath and latex previews of spork
Generates png for both, and overlays
latex above plotmath.
}
\examples{
library(magrittr)
specials <- '& \% $ # \\\\_ { } ~ \\\\^ \\\\'
specials \%>\% as_spork \%>\% as_previews
specials \%>\% gsub(' ','',.) \%>\% as_spork \%>\% as_previews
'one joule (Omega) ~ 1 kg*m^2./s^2' \%>\% as_spork \%>\% as_previews

# possible mismatches between plotmath and latex:

'epsilon.varepsilon' \%>\% as_spork \%>\% as_previews
'rho.varrho' \%>\% as_spork \%>\% as_previews
'Upsilon.Upsilon1' \%>\% as_spork \%>\% as_previews
'phi.phi1.varphi' \%>\% as_spork \%>\% as_previews
'sigma.sigma1.varsigma.stigma' \%>\% as_spork \%>\% as_previews
'theta.vartheta.theta1' \%>\% as_spork \%>\% as_previews
'omega.omega1.pi.varpi' \%>\% as_spork \%>\% as_previews
}
\seealso{
Other preview: 
\code{\link{as.png.plotmath}()},
\code{\link{as.png.spork}()},
\code{\link{as_preview.latex}()},
\code{\link{as_preview.plotmath}()},
\code{\link{as_previews}()},
\code{\link{as_preview}()},
\code{\link{ggplot.plotmath}()},
\code{\link{ggplot.spork}()}
}
\concept{preview}
