% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spork.R
\name{as_spork.character}
\alias{as_spork.character}
\title{Coerce Character to Spork}
\usage{
\method{as_spork}{character}(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{ignored arguments}
}
\value{
spork
}
\description{
Coerces character to class 'spork'.
See description for \code{\link{as_spork}}.
}
\examples{
as_spork('V_c./F')
}
\seealso{
Other spork: 
\code{\link{[.spork}()},
\code{\link{[[.spork}()},
\code{\link{as.list.spork}()},
\code{\link{as.png.spork}()},
\code{\link{as_latex.spork}()},
\code{\link{as_plotmath.spork}()},
\code{\link{as_previews.spork}()},
\code{\link{as_spar.spork}()},
\code{\link{as_spork.factor}()},
\code{\link{as_spork.spork}()},
\code{\link{as_spork}()},
\code{\link{ggplot.spork}()}

Other interface: 
\code{\link{as.expression.plotmath}()},
\code{\link{as_latex.spar}()},
\code{\link{as_latex.spork}()},
\code{\link{as_plotmath.spar}()},
\code{\link{as_plotmath.spork}()},
\code{\link{as_previews.spork}()},
\code{\link{latexToken}()},
\code{\link{plotmathToken}()}

Other character: 
\code{\link{concatenate.character}()}
}
\concept{character}
\concept{interface}
\concept{spork}
