% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate.R
\name{concatenate.character}
\alias{concatenate.character}
\title{Concatenate Character}
\usage{
\method{concatenate}{character}(x, sep = "", ...)
}
\arguments{
\item{x}{character}

\item{sep}{character}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Concatenates Character. Collapses vector using separator.
}
\examples{
concatenate(letters)
}
\seealso{
Other interface: 
\code{\link{as.expression.plotmath}()},
\code{\link{as_latex.spar}()},
\code{\link{as_latex.spork}()},
\code{\link{as_plotmath.spar}()},
\code{\link{as_plotmath.spork}()},
\code{\link{as_spork.character}()},
\code{\link{concatenate.latex}()},
\code{\link{concatenate.plotmath}()},
\code{\link{latexToken}()},
\code{\link{plotmathToken}()}

Other concatenate: 
\code{\link{concatenate.NULL}()},
\code{\link{concatenate.latex}()},
\code{\link{concatenate.plotmath}()},
\code{\link{concatenate}()}
}
\concept{concatenate}
\concept{interface}
