% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate.R
\name{concatenate.plotmath}
\alias{concatenate.plotmath}
\title{Concatenate Plotmath}
\usage{
\method{concatenate}{plotmath}(x, sep = "*", ...)
}
\arguments{
\item{x}{plotmath}

\item{sep}{character}

\item{...}{passed arguments}
}
\value{
plotmath
}
\description{
Concatenates plotmath. Collapses vector using separator.
}
\examples{
concatenate(as_plotmath(as_spork(c('BMI_i','kg/m^2'))))
}
\seealso{
Other concatenate: 
\code{\link{concatenate.NULL}()},
\code{\link{concatenate.character}()},
\code{\link{concatenate.latex}()},
\code{\link{concatenate}()}

Other plotmath: 
\code{\link{[.plotmath}()},
\code{\link{[[.plotmath}()},
\code{\link{as.expression.plotmath}()},
\code{\link{as.png.plotmath}()},
\code{\link{as_plotmath.greek}()},
\code{\link{as_plotmath.spar}()},
\code{\link{as_plotmath.spork}()},
\code{\link{as_plotmath}()},
\code{\link{ggplot.plotmath}()},
\code{\link{goodToken}()},
\code{\link{plotmathToken}()}
}
\concept{concatenate}
\concept{plotmath}
\keyword{internal}
