% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{ggplot.plotmath}
\alias{ggplot.plotmath}
\title{Plot Plotmath}
\usage{
\method{ggplot}{plotmath}(data, mapping = aes(), ...,
  environment = parent.frame(), blank = TRUE)
}
\arguments{
\item{data}{length-one plotmath; see \code{\link{as_plotmath}}}

\item{mapping}{ignored}

\item{...}{ignored arguments}

\item{environment}{ignored}

\item{blank}{whether to use a blank plot area}
}
\value{
gg
}
\description{
Render plotmath in a ggplot.
}
\examples{
library(magrittr)
'one joule (Omega) ~ 1 kg*m^2./s^2' \%>\% as_spork \%>\% as_plotmath \%>\% ggplot
}
\seealso{
Other preview: \code{\link{as.png.plotmath}},
  \code{\link{as.png.spork}},
  \code{\link{as_preview.latex}},
  \code{\link{as_preview.plotmath}},
  \code{\link{as_previews.spork}},
  \code{\link{as_previews}}, \code{\link{as_preview}},
  \code{\link{ggplot.spork}}

Other ggplot: \code{\link{ggplot.spork}}

Other plotmath: \code{\link{[.plotmath}},
  \code{\link{[[.plotmath}},
  \code{\link{as.expression.plotmath}},
  \code{\link{as.png.plotmath}},
  \code{\link{as_plotmath.spar}},
  \code{\link{as_plotmath.spork}},
  \code{\link{as_plotmath}},
  \code{\link{concatenate.plotmath}},
  \code{\link{goodToken}}, \code{\link{plotmathToken}}
}
\concept{ggplot}
\concept{plotmath}
\concept{preview}
\keyword{internal}
