% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{ggplot.spork}
\alias{ggplot.spork}
\title{Plot Spork}
\usage{
\method{ggplot}{spork}(x, blank = TRUE, ...)
}
\arguments{
\item{x}{length-one spork; see \code{\link{as_spork}}}

\item{blank}{whether to use a blank plot area}

\item{...}{ignored arguments}
}
\value{
gg
}
\description{
Render spork in a ggplot.
}
\examples{
'one joule (Omega) ~ 1 kg*m^2./s^2' \%>\% as_spork \%>\% ggplot
}
\seealso{
Other preview: 
\code{\link{as.png.plotmath}()},
\code{\link{as.png.spork}()},
\code{\link{as_preview.latex}()},
\code{\link{as_preview.plotmath}()},
\code{\link{as_previews.spork}()},
\code{\link{as_previews}()},
\code{\link{as_preview}()},
\code{\link{ggplot.plotmath}()}
}
\concept{preview}
\keyword{internal}
