% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmath.R
\name{goodToken}
\alias{goodToken}
\title{Test Whether Token is Parseable}
\usage{
goodToken(x, ...)
}
\arguments{
\item{x}{length-one character}

\item{...}{ignored arguments}
}
\value{
logical
}
\description{
Tests whether token is Parseable
}
\examples{
goodToken('alpha')
goodToken('foo')
goodToken('\\\\$')
}
\seealso{
Other plotmath: \code{\link{[.plotmath}},
  \code{\link{[[.plotmath}},
  \code{\link{as.expression.plotmath}},
  \code{\link{as.png.plotmath}},
  \code{\link{as_plotmath.spar}},
  \code{\link{as_plotmath.spork}},
  \code{\link{as_plotmath}},
  \code{\link{concatenate.plotmath}},
  \code{\link{ggplot.plotmath}},
  \code{\link{plotmathToken}}
}
\concept{plotmath}
\keyword{internal}
