% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTemplate.R
\name{readTemplate}
\alias{readTemplate}
\title{Read in a template file and apply optional replacements}
\usage{
readTemplate(f, addLineBreaks = FALSE, ...)
}
\arguments{
\item{f}{character value; path to template file}

\item{addLineBreaks}{logical value; include line breaks (\code{\\n}) in 
result?}

\item{...}{objects to use as string replacements}
}
\value{
A character value containing the contents of \code{f} after any 
specified string replacements have been made.
}
\description{
\code{readTemplate} reads \code{f} and uses \code{sprintf} to 
make string replacements when specified
}
\examples{
x <- c("select \%s", "from \%s", "where \%s")
write(x, "x.sql")
readTemplate("x.sql", FALSE, "*", "Table", "1=1")
y <- readTemplate("x.sql", TRUE, "*", "Table", "1=1")
print(y)
write(y, "x2.sql")
}
\seealso{
\code{\link{readLines}, \link{sprintf}}
}

