% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spork.R
\name{[.spork}
\alias{[.spork}
\title{Subset Spork}
\usage{
\method{[}{spork}(x, ...)
}
\arguments{
\item{x}{spork}

\item{...}{passed to next method}
}
\value{
spork
}
\description{
Subsets spork, retaining class.
}
\examples{
x <- c(
  'V_c./F',
  'AUC_ss',
  'C_max_ss',
  'var^eta_j'
)
x <- as_spork(x)
class(x)
class(x[1])
}
\seealso{
Other util: \code{\link{[.latex}},
  \code{\link{[.plotmath}}, \code{\link{[[.latex}},
  \code{\link{[[.plotmath}}, \code{\link{[[.spork}}
}
\concept{util}
\keyword{internal}
