% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spork.R
\name{[[.plotmath}
\alias{[[.plotmath}
\title{Element-select Plotmath}
\usage{
\method{[[}{plotmath}(x, ...)
}
\arguments{
\item{x}{plotmath}

\item{...}{passed to next method}
}
\value{
plotmath
}
\description{
Element-selects plotmath, retaining class.
}
\examples{
x <- c(
  'V_c./F',
  'AUC_ss',
  'C_max_ss',
  'var^eta_j'
)
x <- as_plotmath(as_spork(x))
class(x)
class(x[[1]])
}
\seealso{
Other util: \code{\link{[.latex}},
  \code{\link{[.plotmath}}, \code{\link{[.spork}},
  \code{\link{[[.latex}}, \code{\link{[[.spork}}
}
\concept{util}
\keyword{internal}
